+++ src/components/Project/Resource/AddFromServiceDialog.tsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from "react";
-
 import {
     Box,
     Button,
@@ -12,9 +11,15 @@ import {
     Option,
     Select,
     Stack,
-    Typography
-} from "@mui/material";
-import { fetchResource, toSingular } from "@app/utils";
+    Typography,
+    Accordion,
+    AccordionSummary,
+    AccordionDetails,
+    AccordionGroup
+} from "@mui/joy";
+import { downloadMetadata, searchResource, toSingular } from "@app/utils";
+import { NestedInfoTable } from "@app/components/Project/Resource/NestedInfoTable";
+import { Pagination } from "@app/components/Home/Pagination";
 
 interface AddFromServiceDialogProps {
     projectId: string;
@@ -31,60 +36,70 @@ export const AddFromServiceDialog: React.FC<AddFromServiceDialogProps> = ({
     onClose,
     onAddClick
 }) => {
-    const [resourceId, setResourceId] = useState("");
     const [hazardType, setHazardType] = useState("");
-    const [validationMessage, setValidationMessage] = useState<string | null>(null);
-    const [isResourceValid, setIsResourceValid] = useState(false);
+    const [message, setMessage] = useState<string | null>(null);
+    const [messageType, setMessageType] = useState<"success" | "danger" | "primary" | "neutral" | undefined>("danger");
     const [resource, setResource] = useState<Hazard | Dataset | DFR3Mapping | null>(null);
+    const [searchQuery, setSearchQuery] = useState("");
+    const [searchResults, setSearchResults] = useState<any[]>([]);
+    const [expandedAccordion, setExpandedAccordion] = useState<string | null>(null); // Tracks expanded accordion
+
+    // Pagination states
+    const [pageNumber, setPageNumber] = useState(1);
+    const dataPerPage = 10; // Default results per page
+    const nextPage = () => {
+        setPageNumber((prevPage) => prevPage + 1);
+    };
+    const previousPage = () => {
+        setPageNumber((prevPage) => Math.max(prevPage - 1, 1)); // Prevent going below page 1
+    };
 
     useEffect(() => {
-        const validateResource = async () => {
-            if (!resourceId) {
-                setValidationMessage(null);
-                setIsResourceValid(false);
-                setResource(null);
+        setResource(null); // Reset resource when search conditions change
+        const fetchSearchResources = async () => {
+            if (!searchQuery) {
+                setSearchResults([]);
                 return;
             }
 
-            const result = await fetchResource(resourceType, resourceId, hazardType);
-            if (result.error) {
-                setValidationMessage(result.error);
-                setIsResourceValid(false);
-                setResource(null);
+            const results = await searchResource(
+                resourceType,
+                searchQuery,
+                dataPerPage,
+                (pageNumber - 1) * dataPerPage,
+                hazardType
+            );
+            if (results.error) {
+                setMessage(results.error);
+                setMessageType("danger");
+                setSearchResults([]);
             } else {
-                setValidationMessage(null);
-                setIsResourceValid(true);
-
-                // adjust type of the resource
-                if (hazardType) result.type = toSingular(hazardType);
-                if (resourceType.toLowerCase() === "dfr3 mapping") result.type = result.mappingType;
-                setResource(result);
+                setMessage(null);
+                setSearchResults(results);
             }
         };
 
-        validateResource();
-    }, [resourceId, resourceType, hazardType]);
+        fetchSearchResources();
+    }, [searchQuery, resourceType, hazardType, pageNumber]);
+
+    useEffect(() => {
+        setPageNumber(1); // Reset page number when search conditions change
+    }, [searchQuery, resourceType, hazardType]);
 
     return (
         <Modal open={open} onClose={onClose}>
-            <ModalDialog size="lg" sx={{ backgroundColor: "#fff" }}>
-                <ModalClose />
-                <Box
-                    sx={{
-                        width: 600,
-                        maxWidth: "100%",
-                        padding: 2,
-                        borderRadius: "md"
-                    }}
-                >
+            <ModalDialog sx={{ backgroundColor: "#fff", width: "50em" }}>
+                <ModalClose sx={{ zIndex: 20 }} />
+                <Box sx={{ maxWidth: "100%", padding: "5%", overflow: "auto" }}>
                     <Typography level="h4" sx={{ mb: 1, textTransform: "capitalize" }}>
                         Add {resourceType} to Project
                     </Typography>
+
                     {resourceType === "hazard" && (
-                        <FormControl required>
+                        <FormControl required sx={{ marginTop: "1em" }}>
                             <FormLabel>Select Hazard Type</FormLabel>
                             <Select
-                                placeholder="Type"
+                                placeholder="Hazard Type"
                                 value={hazardType}
                                 onChange={(_, newValue) => setHazardType(newValue || "")}
                             >
@@ -96,56 +111,125 @@ export const AddFromServiceDialog: React.FC<AddFromServiceDialogProps> = ({
                             </Select>
                         </FormControl>
                     )}
+
+                    {/* Search Bar */}
+                    <FormControl required sx={{ marginTop: "1em" }}>
+                        <FormLabel>Search by Name or ID</FormLabel>
+                        <Input
+                            placeholder="Name or ID"
+                            value={searchQuery}
+                            onChange={(e) => setSearchQuery(e.target.value)}
+                            disabled={!hazardType && resourceType === "hazard"}
+                        />
+                    </FormControl>
+                    {message && (
+                        <Typography color={messageType} sx={{ mt: 1 }}>
+                            {message}
+                        </Typography>
+                    )}
+
+                    {/* Search Results */}
                     <Stack spacing={2} sx={{ mt: 2 }}>
-                        <FormControl required>
-                            <FormLabel sx={{ textTransform: "capitalize" }}>{resourceType} ID</FormLabel>
-                            <Input
-                                placeholder="ID"
-                                value={resourceId}
-                                onChange={(e) => setResourceId(e.target.value)}
-                                error={!!validationMessage}
-                            />
-                            {validationMessage && (
-                                <Typography color="danger" sx={{ mt: 1 }}>
-                                    {validationMessage}
-                                </Typography>
-                            )}
-                            {isResourceValid && resource && (
-                                <Typography color="success" sx={{ mt: 1 }}>
-                                    {
-                                        // eslint-disable-next-line no-nested-ternary
-                                        "description" in resource && resource.description
-                                            ? resource.description
-                                            : // eslint-disable-next-line no-nested-ternary
-                                              "name" in resource && resource.name
-                                              ? resource.name
-                                              : "title" in resource && resource.title
-                                                ? resource.title
-                                                : "Resource is valid"
-                                    }{" "}
-                                    - {resource.owner}
-                                </Typography>
-                            )}
-                        </FormControl>
+                        <AccordionGroup>
+                            {searchResults.map((result) => {
+                                const isSelected = resource?.id === result.id; // Check if the result is selected
+                                const isExpanded = expandedAccordion === result.id; // Check if the accordion is expanded
+
+                                return (
+                                    <Accordion
+                                        key={result.id}
+                                        expanded={isExpanded} // Expand only the active accordion
+                                        onChange={() =>
+                                            setExpandedAccordion(expandedAccordion === result.id ? null : result.id)
+                                        }
+                                        sx={{ backgroundColor: isExpanded ? "neutral.50" : "inherit" }}
+                                    >
+                                        <AccordionSummary
+                                            sx={{
+                                                color: isSelected ? "primary" : "inherit",
+                                                fontWeight: isSelected ? "bold" : "normal"
+                                            }}
+                                        >
+                                            {result.name ?? result.title ?? "Unnamed Resource"} -{" "}
+                                            {result.owner ?? "Unknown Owner"}
+                                        </AccordionSummary>
+                                        <AccordionDetails>
+                                            {/* Button group on top */}
+                                            <Box
+                                                sx={{
+                                                    display: "flex",
+                                                    justifyContent: "flex-start",
+                                                    gap: 1
+                                                }}
+                                                mt={1}
+                                                mb={1}
+                                            >
+                                                <Button
+                                                    variant="solid"
+                                                    sx={{ backgroundColor: "primary.main" }}
+                                                    onClick={() => {
+                                                        if (hazardType) result.type = toSingular(hazardType);
+                                                        if (resourceType.toLowerCase() === "dfr3 mapping")
+                                                            result.type = result.mappingType;
+                                                        setResource(result);
+                                                        setMessage(
+                                                            `Adding "${
+                                                                result.name || result.title || "Unnamed Resource"
+                                                            }" to project?`
+                                                        );
+                                                        setMessageType("success");
+                                                        setExpandedAccordion(null); // Collapse all accordions
+                                                    }}
+                                                >
+                                                    Select
+                                                </Button>
+                                                <Button
+                                                    variant="outlined"
+                                                    sx={{ color: "primary.main", borderColor: "primary.main" }}
+                                                >
+                                                    Preview
+                                                </Button>
+                                                <Button
+                                                    variant="outlined"
+                                                    sx={{ color: "primary.main", borderColor: "primary.main" }}
+                                                    onClick={() => {
+                                                        downloadMetadata(result);
+                                                    }}
+                                                >
+                                                    Download Metadata
+                                                </Button>
+                                            </Box>
+                                            <NestedInfoTable data={result} />
+                                        </AccordionDetails>
+                                    </Accordion>
+                                );
+                            })}
+                        </AccordionGroup>
+                        {searchResults.length > 0 && (
+                            <Box mt={4} display="flex" justifyContent="center">
+                                <Pagination
+                                    pageNumber={pageNumber}
+                                    data={searchResults}
+                                    dataPerPage={dataPerPage}
+                                    previous={previousPage}
+                                    next={nextPage}
+                                />
+                            </Box>
+                        )}
                     </Stack>
                     <Stack direction="row" spacing={1} sx={{ mt: 3, justifyContent: "flex-end" }}>
-                        <Button variant="plain" onClick={onClose}>
+                        <Button variant="plain" onClick={onClose} sx={{ color: "primary.main" }}>
                             Cancel
                         </Button>
                         <Button
                             variant="solid"
-                            disabled={
-                                resourceType === "hazard"
-                                    ? !resourceId || !hazardType || !isResourceValid
-                                    : !resourceId || !isResourceValid
-                            }
+                            sx={{ backgroundColor: "primary.main" }}
+                            disabled={!resource}
                             onClick={() => {
                                 onAddClick(projectId, resource);
                                 setResource(null);
-                                setResourceId("");
                                 setHazardType("");
-                                setValidationMessage(null);
-                                setIsResourceValid(false);
+                                setMessage(null);
                             }}
                         >
                             Add to Project