+++ src/components/Project/Resource/VisualizationPage.tsx
@@ -51,12 +51,23 @@ const VisualizationPage = (): JSX.Element => {
         }
     }, [id, visualizationPageNumber, deletedVisualizationIds]);
 
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const onSearchClick = () => {};
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const onFilterClick = () => {};
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const onSortClick = () => {};
+    const onApplyFilterSort = (params: { filters: Record<string, string | number>; sortBy: string; order: string }) => {
+        if (id) {
+            const { filters, sortBy, order } = params;
+
+            // Dispatch the Redux Thunk with updated parameters
+            appDispatch(
+                getProjectVisualizations({
+                    projectId: id,
+                    skip: (visualizationPageNumber - 1) * 10, // Pagination logic
+                    limit: 10, // Number of items per page
+                    filters, // Filters applied
+                    sortBy, // Sorting field
+                    order // Sort order: "asc" or "desc"
+                })
+            );
+        }
+    };
 
     // Table view vs Card view
     const [isTableView, setIsTableView] = useState(false); // Toggle state for view mode
@@ -127,13 +138,13 @@ const VisualizationPage = (): JSX.Element => {
                                 <ResourceFilterBar
                                     title="Visualizations"
                                     icon={<VisualizationIcon sx={{ verticalAlign: "middle" }} />}
-                                    onSearchClick={onSearchClick}
-                                    onFilterClick={onFilterClick}
+                                    filters={{ type: ["MAP", "CHART", "TABLE"] }}
+                                    sortOptions={["date", "type", "name", "id"]}
+                                    onApply={onApplyFilterSort}
                                     onCreateClick={onCreateClick}
-                                    onSortClick={onSortClick}
                                     onViewChangeClick={onViewChangeClick}
                                     isTableView={isTableView}
-                                    createLabel="Create New Visualization"
+                                    createLabel="Create Visualization"
                                 />
                                 <CreateVisualizationDialog
                                     projectId={project.id}