+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+### Added
+
+- Checks on the workflow-editor to prevent the user to make edits on finalized workflow. [#73](https://github.com/IN-CORE/incore-studio/issues/73)
+
 ### Fixed
 
 - Stale state cleanup and ensure correctness of the state on page load through internal navigation [#79](https://github.com/IN-CORE/incore-studio/issues/79)
@@ -33,10 +38,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Ability to fill out parameters and input datasets for execution [#61](https://github.com/IN-CORE/incore-studio/issues/61)
 - Display status of each execution steps [#39](https://github.com/IN-CORE/incore-studio/issues/39)
 - Modal to create a new workflow and add to project [#41](https://github.com/IN-CORE/incore-studio/issues/41)
+- Implement search, filter and sort for project resources [#48](https://github.com/IN-CORE/incore-studio/issues/48)
 
 ### Changed
 
 - Navbar file structure. [#37](https://github.com/IN-CORE/incore-studio/issues/37)
+- Improve "add from service" modal. [#66](https://github.com/IN-CORE/incore-studio/issues/66)
 
 ### Fixed